+++ fj-core/src/test/java/test/org/fugerit/java/core/db/dao/daogen/TestBasicDAOHelper.java
@@ -39,7 +39,6 @@ class TestBasicDAOHelper extends TestBasicDBHelper {
 				Assertions.assertNotNull( helper.getLogger() );
 			}
 		} );
-		
 	}
 	
 }