+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- options for statement configuration in BasicDAOHelper <https://github.com/fugerit-org/fj-lib/issues/114>
+
 ## [8.7.2] - 2025-11-20
+
+### Changed
 
 - BasicDAOHelper has method to set query timeout