+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [8.7.3] - 2025-11-22
+
 ### Changed
 
 - options for statement configuration in BasicDAOHelper <https://github.com/fugerit-org/fj-lib/issues/114>